NEWSLETTER/requirements.txt
# YouTube API
google-api-python-client==2.149.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0

# Google Gemini
google-generativeai==0.8.3

# Streamlit
streamlit==1.40.2

# Utilities
python-dotenv==1.0.1
pandas==2.2.3
yt-dlp==2025.11.12

# Date/Time
python-dateutil==2.9.0
